+++ docs/source-app/api_reference/components.rst
@@ -21,5 +21,5 @@ ___________________
     ~python.popen.PopenPythonScript
     ~python.tracer.TracerPythonScript
     ~training.LightningTrainerScript
-    ~serve.gradio.ServeGradio
+    ~serve.gradio_server.ServeGradio
     ~serve.serve.ModelInferenceAPI
