+++ src/lightning_app/components/__init__.py
@@ -9,7 +9,7 @@ from lightning_app.components.multi_node import (
 from lightning_app.components.python.popen import PopenPythonScript
 from lightning_app.components.python.tracer import Code, TracerPythonScript
 from lightning_app.components.serve.auto_scaler import AutoScaler, ColdStartProxy
-from lightning_app.components.serve.gradio import ServeGradio
+from lightning_app.components.serve.gradio_server import ServeGradio
 from lightning_app.components.serve.python_server import Category, Image, Number, PythonServer, Text
 from lightning_app.components.serve.serve import ModelInferenceAPI
 from lightning_app.components.serve.streamlit import ServeStreamlit
