+++ src/lightning_app/components/serve/__init__.py
@@ -1,5 +1,5 @@
 from lightning_app.components.serve.auto_scaler import AutoScaler, ColdStartProxy
-from lightning_app.components.serve.gradio import ServeGradio
+from lightning_app.components.serve.gradio_server import ServeGradio
 from lightning_app.components.serve.python_server import Category, Image, Number, PythonServer, Text
 from lightning_app.components.serve.streamlit import ServeStreamlit
